Get rid of old DBServerEdgeCursor
