--- arangod/Graph/Cursors/DBServerEdgeCursor.cpp
@@ -1,421 +1,0 @@
-////////////////////////////////////////////////////////////////////////////////
-/// DISCLAIMER
-///
-/// Copyright 2014-2024 ArangoDB GmbH, Cologne, Germany
-/// Copyright 2004-2014 triAGENS GmbH, Cologne, Germany
-///
-/// Licensed under the Business Source License 1.1 (the "License");
-/// you may not use this file except in compliance with the License.
-/// You may obtain a copy of the License at
-///
-///     https://github.com/arangodb/arangodb/blob/devel/LICENSE
-///
-/// Unless required by applicable law or agreed to in writing, software
-/// distributed under the License is distributed on an "AS IS" BASIS,
-/// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-/// See the License for the specific language governing permissions and
-/// limitations under the License.
-///
-/// Copyright holder is ArangoDB GmbH, Cologne, Germany
-///
-/// @author Michael Hackstein
-////////////////////////////////////////////////////////////////////////////////
-
-#include "DBServerEdgeCursor.h"
-
-#include "Aql/AstNode.h"
-#include "Aql/AttributeNamePath.h"
-#include "Aql/QueryContext.h"
-#include "Aql/Projections.h"
-#include "Basics/StaticStrings.h"
-#include "Graph/BaseOptions.h"
-#include "Graph/EdgeDocumentToken.h"
-#include "Graph/TraverserCache.h"
-#include "Indexes/IndexIterator.h"
-#include "StorageEngine/PhysicalCollection.h"
-#include "StorageEngine/TransactionState.h"
-#include "Transaction/Helpers.h"
-#include "Transaction/Methods.h"
-#include "VocBase/Identifiers/DataSourceId.h"
-#include "VocBase/LogicalCollection.h"
-
-using namespace arangodb;
-using namespace arangodb::graph;
-
-namespace {
-void PrepareIndexCondition(BaseOptions::LookupInfo const& info,
-                           std::string_view vertex) {
-  auto& node = info.indexCondition;
-  TRI_ASSERT(node->numMembers() > 0);
-  if (info.conditionNeedUpdate) {
-    // We have to inject _from/_to iff the condition needs it
-    auto dirCmp = node->getMemberUnchecked(info.conditionMemberToUpdate);
-    TRI_ASSERT(dirCmp->type == aql::NODE_TYPE_OPERATOR_BINARY_EQ);
-    TRI_ASSERT(dirCmp->numMembers() == 2);
-
-    auto idNode = dirCmp->getMemberUnchecked(1);
-    TRI_ASSERT(idNode->type == aql::NODE_TYPE_VALUE);
-    TRI_ASSERT(idNode->isValueType(aql::VALUE_TYPE_STRING));
-    // must edit node in place; TODO replace node?
-    TEMPORARILY_UNLOCK_NODE(idNode);
-    idNode->setStringValue(vertex.data(), vertex.length());
-  }
-}
-}  // namespace
-
-////////////////////////////////////////////////////////////////////////////////
-/// @brief Get a document by it's ID. Also lazy locks the collection.
-///        If DOCUMENT_NOT_FOUND this function will return normally
-///        with a OperationResult.failed() == true.
-///        On all other cases this function throws.
-////////////////////////////////////////////////////////////////////////////////
-
-DBServerEdgeCursor::DBServerEdgeCursor(
-    BaseOptions* opts, aql::Variable const* tmpVar,
-    std::vector<BaseOptions::LookupInfo> const& lookupInfo)
-    : _opts(opts),
-      _monitor(_opts->query().resourceMonitor()),
-      _trx(opts->trx()),
-      _tmpVar(tmpVar),
-      _currentCursor(0),
-      _currentSubCursor(0),
-      _cachePos(0),
-      _lookupInfo(lookupInfo) {
-  _cache.reserve(1000);
-  TRI_ASSERT(_opts->cache() != nullptr);
-
-  if (_opts->cache() == nullptr) {
-    THROW_ARANGO_EXCEPTION_MESSAGE(
-        TRI_ERROR_INTERNAL, "no cache present for single server edge cursor");
-  }
-}
-
-DBServerEdgeCursor::~DBServerEdgeCursor() = default;
-
-#ifdef USE_ENTERPRISE
-static bool CheckInaccessible(transaction::Methods* trx, VPackSlice edge) {
-  // for skipInaccessibleCollections we need to check the edge
-  // document, in that case nextWithExtra has no benefit
-  TRI_ASSERT(edge.isString());
-  std::string_view str(edge.stringView());
-  size_t pos = str.find('/');
-  TRI_ASSERT(pos != std::string_view::npos);
-  return trx->isInaccessibleCollection(str.substr(0, pos));
-}
-#endif
-
-void DBServerEdgeCursor::getDocAndRunCallback(
-    IndexIterator* cursor, EdgeCursor::Callback const& callback) {
-  auto collection = cursor->collection();
-  EdgeDocumentToken etkn(collection->id(), _cache[_cachePos++]);
-  auto cb = [&](LocalDocumentId, aql::DocumentData&&, VPackSlice edgeDoc) {
-#ifdef USE_ENTERPRISE
-    if (_trx->skipInaccessible()) {
-      // TODO: we only need to check one of these
-      VPackSlice from = transaction::helpers::extractFromFromDocument(edgeDoc);
-      VPackSlice to = transaction::helpers::extractToFromDocument(edgeDoc);
-      if (CheckInaccessible(_trx, from) || CheckInaccessible(_trx, to)) {
-        return false;
-      }
-    }
-#endif
-    _opts->cache()->incrDocuments();
-    callback(std::move(etkn), edgeDoc, _currentCursor);
-    return true;
-  };
-  collection->getPhysical()->lookup(_trx, etkn.localDocumentId(), cb,
-                                    {.countBytes = true});
-}
-std::function<bool(LocalDocumentId, aql::DocumentData&&, VPackSlice)>
-DBServerEdgeCursor::nonCoveringCallback(DataSourceId const& sourceId,
-                                        size_t cursorId,
-                                        EdgeCursor::Callback const& callback) {
-  return [=](LocalDocumentId token, aql::DocumentData&&, VPackSlice edgeDoc) {
-#ifdef USE_ENTERPRISE
-    if (_trx->skipInaccessible()) {
-      // TODO: we only need to check one of these
-      VPackSlice from = transaction::helpers::extractFromFromDocument(edgeDoc);
-      VPackSlice to = transaction::helpers::extractToFromDocument(edgeDoc);
-      if (CheckInaccessible(_trx, from) || CheckInaccessible(_trx, to)) {
-        return false;
-      }
-    }
-#endif
-    _opts->cache()->incrDocuments();
-    callback(EdgeDocumentToken(sourceId, token), edgeDoc, cursorId);
-    return true;
-  };
-}
-std::function<bool(LocalDocumentId, IndexIteratorCoveringData&)>
-DBServerEdgeCursor::coveringCallback(bool& operationSuccessful,
-                                     DataSourceId const& sourceId,
-                                     size_t cursorId, uint16_t coveringPosition,
-                                     EdgeCursor::Callback const& callback) {
-  return [=, &operationSuccessful](LocalDocumentId token,
-                                   IndexIteratorCoveringData& covering) {
-    TRI_ASSERT(covering.isArray());
-    VPackSlice edge = covering.at(coveringPosition);
-    TRI_ASSERT(edge.isString());
-
-    if (token.isSet()) {
-#ifdef USE_ENTERPRISE
-      if (_trx->skipInaccessible() && CheckInaccessible(_trx, edge)) {
-        return false;
-      }
-#endif
-      operationSuccessful = true;
-      _opts->cache()->incrDocuments();
-      auto etkn = EdgeDocumentToken(sourceId, token);
-      callback(EdgeDocumentToken(sourceId, token), edge, cursorId);
-      return true;
-    }
-    return false;
-  };
-}
-
-bool DBServerEdgeCursor::advanceCursor(IndexIterator*& cursor,
-                                       std::vector<CursorInfo>*& cursorSet) {
-  TRI_ASSERT(!_cursors.empty());
-  ++_currentSubCursor;
-  if (_currentSubCursor >= cursorSet->size()) {
-    ++_currentCursor;
-    _currentSubCursor = 0;
-    if (_currentCursor == _cursors.size()) {
-      // We are done, all cursors exhausted.
-      return false;
-    }
-    cursorSet = &_cursors[_currentCursor];
-  }
-
-  cursor = (*cursorSet)[_currentSubCursor].cursor.get();
-  // If we switch the cursor. We have to clear the cache.
-  _cache.clear();
-  return true;
-}
-
-bool DBServerEdgeCursor::next(EdgeCursor::Callback const& callback) {
-  // fills callback with next EdgeDocumentToken and Slice that contains the
-  // other side of the edge (we are standing on a node and want to iterate all
-  // connected edges
-  TRI_ASSERT(!_cursors.empty());
-
-  // curiously enough, this method is only called in the cluster, but not
-  // on single servers.
-
-  if (_currentCursor == _cursors.size()) {
-    return false;
-  }
-
-  // get current cursor
-  auto* cursorSet = &_cursors[_currentCursor];
-  auto cursor = (*cursorSet)[_currentSubCursor].cursor.get();
-  uint16_t coveringPosition =
-      (*cursorSet)[_currentSubCursor].coveringIndexPosition;
-  TRI_ASSERT(cursor != nullptr);
-
-  // There is still something in the cache
-  if (_cachePos < _cache.size()) {
-    // get the collection
-    auto cb = nonCoveringCallback(cursor->collection()->id(), _currentCursor,
-                                  callback);
-    cursor->collection()->getPhysical()->lookup(_trx, _cache[_cachePos++], cb,
-                                                {.countBytes = true});
-    return true;
-  }
-
-  // We need to refill the cache.
-  _cachePos = 0;
-  // NOTE: We cannot clear the cache,
-  // because the cursor expects it to be filled.
-
-  if (aql::Projections::isCoveringIndexPosition(coveringPosition)) {
-    bool operationSuccessful = false;
-    do {
-      if (cursorSet->empty() || !cursor->hasMore()) {
-        if (!advanceCursor(cursor, cursorSet)) {
-          return false;
-        }
-      } else {
-        auto cb =
-            coveringCallback(operationSuccessful, cursor->collection()->id(),
-                             _currentCursor, coveringPosition, callback);
-        cursor->nextCovering(cb, 1);
-        if (operationSuccessful) {
-          return true;
-        }
-      }
-    } while (!operationSuccessful);
-  }
-
-  TRI_ASSERT(!aql::Projections::isCoveringIndexPosition(coveringPosition));
-
-  do {
-    if (cursorSet->empty() || !cursor->hasMore()) {
-      if (!advanceCursor(cursor, cursorSet)) {
-        return false;
-      }
-    } else {
-      _cache.clear();
-      bool tmp = cursor->next(
-          [&](LocalDocumentId token) {
-            if (token.isSet()) {
-              _cache.emplace_back(token);
-              return true;
-            }
-            return false;
-          },
-          1000);
-      TRI_ASSERT(tmp == cursor->hasMore());
-    }
-  } while (_cache.empty());
-  TRI_ASSERT(!_cache.empty());
-  TRI_ASSERT(_cachePos < _cache.size());
-  auto cb =
-      nonCoveringCallback(cursor->collection()->id(), _currentCursor, callback);
-  cursor->collection()->getPhysical()->lookup(_trx, _cache[_cachePos++], cb,
-                                              {.countBytes = true});
-  return true;
-}
-
-void DBServerEdgeCursor::readAll(EdgeCursor::Callback const& callback) {
-  TRI_ASSERT(!_cursors.empty());
-  size_t cursorId = 0;
-  for (_currentCursor = 0; _currentCursor < _cursors.size(); ++_currentCursor) {
-    cursorId = _currentCursor;
-    auto& cursorSet = _cursors[_currentCursor];
-    for (auto& [cursor, coveringPosition] : cursorSet) {
-      LogicalCollection* collection = cursor->collection();
-      auto cid = collection->id();
-
-      if (aql::Projections::isCoveringIndexPosition(coveringPosition)) {
-        bool operationSuccessful = true;
-
-        auto cb = coveringCallback(operationSuccessful, cid, cursorId,
-                                   coveringPosition, callback);
-        cursor->allCovering(cb);
-      } else {
-        auto cb = nonCoveringCallback(cid, cursorId, callback);
-        cursor->all([&](LocalDocumentId token) {
-          return collection->getPhysical()
-              ->lookup(_trx, token, cb, {.countBytes = true})
-              .ok();
-        });
-      }
-
-      // update cache hits and misses
-      auto [ch, cm] = cursor->getAndResetCacheStats();
-      _opts->cache()->incrCacheHits(ch);
-      _opts->cache()->incrCacheMisses(cm);
-    }
-  }
-}
-
-void DBServerEdgeCursor::rearm(std::string_view vertex, uint64_t /*depth*/) {
-  _currentCursor = 0;
-  _currentSubCursor = 0;
-  _cache.clear();
-  _cachePos = 0;
-
-  if (_cursors.empty()) {
-    buildLookupInfo(vertex);
-    return;
-  }
-
-  size_t i = 0;
-  for (auto& info : _lookupInfo) {
-    ::PrepareIndexCondition(info, vertex);
-    IndexIteratorOptions defaultIndexIteratorOptions;
-
-    auto& csrs = _cursors[i++];
-    size_t j = 0;
-    auto& node = info.indexCondition;
-    for (auto const& it : info.idxHandles) {
-      auto& cursor = csrs[j].cursor;
-
-      // steal cache hits and misses before the cursor is recycled
-      auto [ch, cm] = cursor->getAndResetCacheStats();
-      _opts->cache()->incrCacheHits(ch);
-      _opts->cache()->incrCacheMisses(cm);
-
-      // check if the underlying index iterator supports rearming
-      if (cursor->canRearm()) {
-        // rearming supported
-        _opts->cache()->incrCursorsRearmed();
-        if (!cursor->rearm(node, _tmpVar, defaultIndexIteratorOptions)) {
-          cursor =
-              std::make_unique<EmptyIndexIterator>(cursor->collection(), _trx);
-        }
-      } else {
-        // rearming not supported - we need to throw away the index iterator
-        // and create a new one
-        _opts->cache()->incrCursorsCreated();
-        cursor = _trx->indexScanForCondition(
-            _monitor, it, node, _tmpVar, defaultIndexIteratorOptions,
-            ReadOwnWrites::no,
-            static_cast<int>(
-                info.conditionNeedUpdate
-                    ? info.conditionMemberToUpdate
-                    : transaction::Methods::kNoMutableConditionIdx));
-      }
-      ++j;
-    }
-  }
-}
-
-void DBServerEdgeCursor::buildLookupInfo(std::string_view vertex) {
-  TRI_ASSERT(_cursors.empty());
-  _cursors.reserve(_lookupInfo.size());
-
-  for (auto& info : _lookupInfo) {
-    addCursor(info, vertex);
-  }
-}
-
-void DBServerEdgeCursor::addCursor(BaseOptions::LookupInfo const& info,
-                                   std::string_view vertex) {
-  ::PrepareIndexCondition(info, vertex);
-  IndexIteratorOptions defaultIndexIteratorOptions;
-
-  _cursors.emplace_back();
-  auto& csrs = _cursors.back();
-  csrs.reserve(info.idxHandles.size());
-  for (std::shared_ptr<Index> const& index : info.idxHandles) {
-    uint16_t coveringPosition = aql::Projections::kNoCoveringIndexPosition;
-
-    // projections we want to cover
-    std::vector<aql::AttributeNamePath> paths = {};
-    paths.emplace_back(
-        aql::AttributeNamePath({StaticStrings::FromString}, _monitor));
-    paths.emplace_back(
-        aql::AttributeNamePath({StaticStrings::ToString}, _monitor));
-    aql::Projections edgeProjections(std::move(paths));
-
-    if (index->covers(edgeProjections)) {
-      // find opposite attribute
-      edgeProjections.setCoveringContext(index->collection().id(), index);
-
-      TRI_edge_direction_e dir = info.direction;
-      TRI_ASSERT(dir == TRI_EDGE_IN || dir == TRI_EDGE_OUT);
-
-      if (dir == TRI_EDGE_OUT) {
-        coveringPosition = edgeProjections.coveringIndexPosition(
-            aql::AttributeNamePath::Type::ToAttribute);
-      } else {
-        coveringPosition = edgeProjections.coveringIndexPosition(
-            aql::AttributeNamePath::Type::FromAttribute);
-      }
-
-      TRI_ASSERT(aql::Projections::isCoveringIndexPosition(coveringPosition));
-    }
-
-    csrs.emplace_back(
-        _trx->indexScanForCondition(
-            _monitor, index, info.indexCondition, _tmpVar,
-            defaultIndexIteratorOptions, ReadOwnWrites::no,
-            static_cast<int>(
-                info.conditionNeedUpdate
-                    ? info.conditionMemberToUpdate
-                    : transaction::Methods::kNoMutableConditionIdx)),
-        coveringPosition);
-  }
-}
