+++ arangod/Graph/TraverserOptions.cpp
@@ -32,7 +32,8 @@
 #include "Basics/VelocyPackHelper.h"
 #include "Basics/tryEmplaceHelper.h"
 #include "Cluster/ClusterEdgeCursor.h"
-#include "Graph/Cursors/NewDBServerEdgeCursor.h"
+#include "Graph/Cursors/DBServerEdgeCursor.h"
+#include "Graph/Cursors/DBServerIndexCursor.h"
 #include "Indexes/Index.h"
 
 #include <velocypack/Iterator.h>
@@ -762,17 +763,16 @@ std::unique_ptr<EdgeCursor> arangodb::traverser::TraverserOptions::buildCursor(
   auto specific = _depthLookupInfo.find(depth);
   if (specific != _depthLookupInfo.end()) {
     return std::make_unique<
-        graph::DBServerEdgeCursor<graph::CollectionIndexCursor>>(
-        graph::createCollectionIndexCursors(specific->second, _tmpVar, trx(),
-                                            cache(),
-                                            query().resourceMonitor()));
+        graph::DBServerEdgeCursor<graph::DBServerIndexCursor>>(
+        graph::createDBServerIndexCursors(specific->second, _tmpVar, trx(),
+                                          cache(), query().resourceMonitor()));
   }
 
   // otherwise, retain / reuse the general (global) cursor
   return std::make_unique<
-      graph::DBServerEdgeCursor<graph::CollectionIndexCursor>>(
-      graph::createCollectionIndexCursors(_baseLookupInfos, _tmpVar, trx(),
-                                          cache(), query().resourceMonitor()));
+      graph::DBServerEdgeCursor<graph::DBServerIndexCursor>>(
+      graph::createDBServerIndexCursors(_baseLookupInfos, _tmpVar, trx(),
+                                        cache(), query().resourceMonitor()));
 }
 
 double TraverserOptions::estimateCost(size_t& nrItems) const {
